Adding customers improvement
	- Customer's name input will be all upper case to reduce ambiguity such
as John and john will not being the same. It will now print JOHN and
JOHN. Changing john to John and leaving John as John is too much coding
for this purpose. (Can do that if you want) I think all caps is fine.

	- Checks the CustomerList for the new Customer being added. This is
done by comparing Name AND Phone Number. If there is an identical match,
error will display and customer will not be added. User can still decide
to add another user.

	- Customer ID will generate AFTER uniqueness test is passed. Otherwise
ID will increment on failed adds.

	- Added a getList() method to the ItemList Class to return the list.

	- Changed toString() in Customer. The output is now indented like
columns for easy visibility. This prints for the collection call as
well. It now prints ID -> Name -> Phone #