Adding washers improvement
	- Adding washers will also check for pre-existing Model and Brand in
the list.
	- If current Model and Brand already exists. It does a comparison on
price values. If the new value is different than what is in the list.
The new value of the washer will be override the old. If the values are
the same error will tell user that washer already exists.

	- Changed toString() format to print in column like output

	- Mutate input strings of addWasher and addToInventory, to change
string to upper case and trim away excess spaces. This makes searching
for Kenmore and kenmore the same.

	- Updated total sales output to 2 decimal $0.00 output. Before it was
$0.0

	- Added comments to methods I added